green: View full details button on map summary sheet opens full detail card (tc-1q07)
Adds the "View full details" PrimaryButton to ApplicationSummarySheet, the
MapView sheet onDismiss that presents the pending detail after the summary
dismisses, and the AppCoordinator.makeMapViewModel wiring. Relocates the
detail-navigation methods to a new AppCoordinator+Detail extension to keep
AppCoordinator under the 400-line SwiftLint limit.

Release-Note: On the map, tap a pin then "View full details" to open the full application and its council portal link.
